third_party/blink/web_tests/fast/events/wheel/mouse-wheel-scroll-latching.html

<!DOCTYPE HTML>
<script src='../../../resources/gesture-util.js'></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-actions.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<style>
  ::-webkit-scrollbar {
    display: none;
  }
  body {
    margin: 0px;
    height: 100px;
    width: 100px;
  }
  #parentDiv {
    background-color: #FF7F7F;
    height: 60px;
    width: 60px;
    overflow: scroll;
  }
  #content1 {
    height: 70px;
    width: 70px;
  }
  #childDiv {
    background-color: #84BE6A;
    height: 50px;
    width: 50px;
    overflow: scroll;
  }
  #content2 {
    height: 60px;
    width: 60px;
  }
</style>

<div id="parentDiv">
  <div id="content1">
    <div id="childDiv">
      <div id="content2">
      </div>
    </div>
  </div>
</div>

<script>
const parentDiv = document.getElementById('parentDiv');
const childDiv = document.getElementById('childDiv');
const pos = elementCenter(childDiv);
const deltaY = 500;

promise_test(async () => {
  await wheelScroll(pos.x, pos.y, 0, deltaY, childDiv);
  assert_equals(childDiv.scrollTop, 10, "childDiv must be fully scrolled");
  await conditionHolds(() => { return parentDiv.scrollTop == 0; },
                       "parentDiv shouldn't scroll at all");
}, 'With wheel scroll latching enabled only child div must scroll');
</script>
